Set up 'memory' and 'metrics' modules
