+++ src/memory.rs
@@ -1,0 +1,51 @@
+use candid::{candid_method, CandidType, Principal};
+use ic_canister_log::{declare_log_buffer, log};
+use ic_canisters_http_types::{
+    HttpRequest as AssetHttpRequest, HttpResponse as AssetHttpResponse, HttpResponseBuilder,
+};
+use ic_cdk::api::management_canister::http_request::{
+    http_request as make_http_request, CanisterHttpRequestArgument, HttpHeader, HttpMethod,
+    HttpResponse, TransformArgs, TransformContext,
+};
+use ic_cdk::{query, update};
+use ic_nervous_system_common::{serve_logs, serve_logs_v2, serve_metrics};
+#[cfg(not(target_arch = "wasm32"))]
+use ic_stable_structures::file_mem::FileMemory;
+use ic_stable_structures::memory_manager::{MemoryId, MemoryManager, VirtualMemory};
+#[cfg(target_arch = "wasm32")]
+use ic_stable_structures::DefaultMemoryImpl;
+use ic_stable_structures::{Cell, StableBTreeMap};
+use std::cell::RefCell;
+use std::collections::hash_set::HashSet;
+
+use crate::types::*;
+
+#[cfg(not(target_arch = "wasm32"))]
+type Memory = VirtualMemory<FileMemory>;
+#[cfg(target_arch = "wasm32")]
+type Memory = VirtualMemory<DefaultMemoryImpl>;
+
+declare_log_buffer!(name = INFO, capacity = 1000);
+declare_log_buffer!(name = ERROR, capacity = 1000);
+
+thread_local! {
+    // Transient static data: this is reset when the canister is upgraded.
+    pub static METRICS: RefCell<Metrics> = RefCell::new(Metrics::default());
+    pub static SERVICE_HOSTS_ALLOWLIST: RefCell<AllowlistSet> = RefCell::new(AllowlistSet::new());
+    pub static AUTH_STABLE: RefCell<HashSet<Principal>> = RefCell::new(HashSet::<Principal>::new());
+
+    // Stable static data: this is preserved when the canister is upgraded.
+    #[cfg(not(target_arch = "wasm32"))]
+    pub static MEMORY_MANAGER: RefCell<MemoryManager<FileMemory>> =
+        RefCell::new(MemoryManager::init(FileMemory::new(std::fs::OpenOptions::new().read(true).write(true).create(true).open("target/test_stable_memory.bin").unwrap())));
+    #[cfg(target_arch = "wasm32")]
+    pub static MEMORY_MANAGER: RefCell<MemoryManager<DefaultMemoryImpl>> =
+        RefCell::new(MemoryManager::init(DefaultMemoryImpl::default()));
+    pub static METADATA: RefCell<Cell<Metadata, Memory>> = RefCell::new(Cell::init(
+            MEMORY_MANAGER.with(|m| m.borrow().get(MemoryId::new(0))),
+            <Metadata>::default()).unwrap());
+    pub static AUTH: RefCell<StableBTreeMap<PrincipalStorable, u32, Memory>> = RefCell::new(
+        StableBTreeMap::init(MEMORY_MANAGER.with(|m| m.borrow().get(MemoryId::new(1)))));
+    pub static PROVIDERS: RefCell<StableBTreeMap<u64, Provider, Memory>> = RefCell::new(
+        StableBTreeMap::init(MEMORY_MANAGER.with(|m| m.borrow().get(MemoryId::new(2)))));
+}
